Add support for Python 3.13+.
This adds support for python3.13+ for the C code so far.

The zipextimporter may or may not work as is for these versions of Python though and could be replaced with these 2 files: https://github.com/SeaHOH/memimport/blob/main/zipextimporter.py and https://github.com/SeaHOH/memimport/blob/main/memimport.py.